refactor: change node exports from /cli to /node
