+++ packages/world/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["mud.config.ts", "ts/library/index.ts", "ts/register/index.ts", "ts/cli/index.ts"],
+  entry: ["mud.config.ts", "ts/library/index.ts", "ts/register/index.ts", "ts/node/index.ts"],
   target: "esnext",
   format: ["esm"],
   dts: false,
